refactoring parser
