+++ helix-db/src/helixc/analyzer/errors.rs
@@ -5,7 +5,7 @@ use crate::helixc::{
         error_codes::ErrorCode,
         fix::Fix,
     },
-    parser::{helix_parser::Query, location::Loc},
+    parser::{location::Loc, types::Query},
 };
 
 pub(crate) fn push_schema_err(
