+++ helix-db/src/helixc/analyzer/methods/infer_expr_type.rs
@@ -1,5 +1,6 @@
 //! Semantic analyzer for Helix‑QL.
 use crate::helixc::analyzer::error_codes::ErrorCode;
+use crate::helixc::analyzer::utils::DEFAULT_VAR_NAME;
 use crate::helixc::generator::utils::EmbedData;
 use crate::{
     generate_error,
@@ -26,7 +27,7 @@ use crate::{
             },
             utils::{GenRef, GeneratedValue, Separator, VecData},
         },
-        parser::helix_parser::*,
+        parser::types::*,
     },
     protocol::date::Date,
 };
@@ -57,7 +58,6 @@ pub(crate) fn infer_expr_type<'a>(
     parent_ty: Option<Type>,
     gen_query: &mut GeneratedQuery,
 ) -> (Type, Option<GeneratedStatement>) {
-    // TODO: Look at returning statement as well or passing mut query to push to
     use ExpressionType::*;
     let expr: &ExpressionType = &expression.expr;
     match expr {
@@ -98,32 +98,40 @@ pub(crate) fn infer_expr_type<'a>(
             Type::Boolean,
             Some(GeneratedStatement::Literal(GenRef::Literal(b.to_string()))),
         ),
+        // Gets expression type for each element in the array
+        // Checks if all elements are of the same type
+        // Returns the type of the array and the statements to generate from the array
         ArrayLiteral(a) => {
             let mut inner_array_ty = None;
-            let stmts = a
-                .iter()
-                .map(|e| {
-                    let (ty, stmt) = infer_expr_type(
-                        ctx,
-                        e,
-                        scope,
-                        original_query,
-                        parent_ty.clone(),
-                        gen_query,
-                    );
-                    if inner_array_ty.is_none() {
-                        inner_array_ty = Some(ty);
-                    } else {
-                        // TODO handle type is same for all elements
+            let result = a.iter().try_fold(Vec::new(), |mut stmts, e| {
+                let (ty, stmt) =
+                    infer_expr_type(ctx, e, scope, original_query, parent_ty.clone(), gen_query);
+                let type_str = ty.kind_str();
+                if let Some(inner_array_ty) = &inner_array_ty {
+                    if inner_array_ty != &ty {
+                        generate_error!(ctx, original_query, e.loc.clone(), E306, type_str);
                     }
-                    // TODO handle none for stmt
-                    stmt.unwrap()
-                })
-                .collect::<Vec<_>>();
-            (
-                inner_array_ty.unwrap(),
-                Some(GeneratedStatement::Array(stmts)),
-            )
+                } else {
+                    inner_array_ty = Some(ty);
+                }
+                match stmt {
+                    Some(s) => {
+                        stmts.push(s);
+                        Ok(stmts)
+                    }
+                    None => {
+                        generate_error!(ctx, original_query, e.loc.clone(), E306, type_str);
+                        Err(())
+                    }
+                }
+            });
+            match result {
+                Ok(stmts) => (
+                    Type::Array(Box::new(inner_array_ty.unwrap())),
+                    Some(GeneratedStatement::Array(stmts)),
+                ),
+                Err(()) => (Type::Unknown, Some(GeneratedStatement::Empty)),
+            }
         }
         Traversal(tr) => {
             let mut gen_traversal = GeneratedTraversal::default();
@@ -136,7 +144,6 @@ pub(crate) fn infer_expr_type<'a>(
                 &mut gen_traversal,
                 gen_query,
             );
-            // push query
             let stmt = GeneratedStatement::Traversal(gen_traversal);
 
             if matches!(expr, Exists(_)) {
@@ -167,14 +174,13 @@ pub(crate) fn infer_expr_type<'a>(
                     .filter_map(|p| p.default_value.clone().map(|v| (p.name.clone(), v)))
                     .collect::<Vec<(String, GeneratedValue)>>();
 
-                // Validate fields if both type and fields are present
+                // Validate fields of add node by traversing the fields
+                // checking they exist in the schema, then checking their types
                 let (properties, secondary_indices) = match &add.fields {
-                    Some(fields) => {
-                        // Get the field set before validation
-                        // TODO: Check field types
-                        let field_set = ctx.node_fields.get(ty.as_str()).cloned();
-                        if let Some(field_set) = field_set {
-                            for (field_name, value) in fields {
+                    Some(fields_to_add) => {
+                        let field_set_from_schema = ctx.node_fields.get(ty.as_str()).cloned();
+                        if let Some(field_set) = field_set_from_schema {
+                            for (field_name, field_value) in fields_to_add {
                                 if !field_set.contains_key(field_name.as_str()) {
                                     generate_error!(
                                         ctx,
@@ -186,7 +192,7 @@ pub(crate) fn infer_expr_type<'a>(
                                         ty.as_str()
                                     );
                                 }
-                                match value {
+                                match field_value {
                                     ValueType::Identifier { value, loc } => {
                                         if is_valid_identifier(
                                             ctx,
@@ -202,10 +208,36 @@ pub(crate) fn infer_expr_type<'a>(
                                                 E301,
                                                 value.as_str()
                                             );
-                                        };
+                                        } else {
+                                            let variable_type =
+                                                scope.get(value.as_str()).unwrap();
+                                            if variable_type
+                                                != &Type::from(
+                                                    field_set
+                                                        .get(field_name.as_str())
+                                                        .unwrap()
+                                                        .field_type
+                                                        .clone(),
+                                                )
+                                            {
+                                                generate_error!(
+                                                    ctx,    
+                                                    original_query,
+                                                    loc.clone(),
+                                                    E205,
+                                                    value.as_str(),
+                                                    &field_set
+                                                        .get(field_name.as_str())
+                                                        .unwrap()
+                                                        .field_type
+                                                        .to_string(),
+                                                    "node",
+                                                    ty.as_str()
+                                                );
+                                            }
+                                        }
                                     }
                                     ValueType::Literal { value, loc } => {
-                                        // check against type
                                         let field_type = ctx
                                             .node_fields
                                             .get(ty.as_str())
@@ -231,7 +263,7 @@ pub(crate) fn infer_expr_type<'a>(
                                 }
                             }
                         }
-                        let mut properties = fields
+                        let mut properties = fields_to_add
                             .iter()
                             .map(|(field_name, value)| {
                                 (
@@ -267,13 +299,7 @@ pub(crate) fn infer_expr_type<'a>(
                                                 )),
                                             }
                                         }
-                                        ValueType::Identifier { value, loc } => {
-                                            is_valid_identifier(
-                                                ctx,
-                                                original_query,
-                                                loc.clone(),
-                                                value.as_str(),
-                                            );
+                                        ValueType::Identifier { value, .. } => {
                                             gen_identifier_or_param(
                                                 original_query,
                                                 value,
@@ -1056,10 +1082,9 @@ pub(crate) fn infer_expr_type<'a>(
             assert!(matches!(stmt, Some(GeneratedStatement::Traversal(_))));
             let traversal = match stmt.unwrap() {
                 GeneratedStatement::Traversal(mut tr) => {
-                    // TODO: FIX VALUE HERE
                     let source_variable = match tr.source_step.inner() {
                         SourceStep::Identifier(id) => id.inner().clone(),
-                        _ => "val".to_string(),
+                        _ => DEFAULT_VAR_NAME.to_string(),
                     };
                     tr.traversal_type = TraversalType::NestedFrom(GenRef::Std(source_variable));
                     tr.should_collect = ShouldCollect::No;
