+++ helix-db/src/helixc/analyzer/methods/migration_validation.rs
@@ -8,7 +8,7 @@ use crate::{
             },
             utils::{GenRef, GeneratedValue, Separator},
         },
-        parser::helix_parser::{
+        parser::types::{
             FieldValueType, Migration, MigrationItem, MigrationPropertyMapping,
         },
     },
