+++ helix-db/src/helixc/analyzer/methods/statement_validation.rs
@@ -16,7 +16,7 @@ use crate::{
             },
             utils::GenRef,
         },
-        parser::helix_parser::*,
+        parser::types::*,
     },
 };
 use paste::paste;
