+++ helix-db/src/helixc/generator/utils.rs
@@ -1,6 +1,5 @@
+use crate::helixc::parser::types::IdType;
 use std::fmt::{self, Debug, Display};
-
-use crate::helixc::parser::helix_parser::IdType;
 
 #[derive(Clone)]
 pub enum GenRef<T>
@@ -298,7 +297,7 @@ impl Display for RustType {
             RustType::F32 => write!(f, "f32"),
             RustType::F64 => write!(f, "f64"),
             RustType::Bool => write!(f, "bool"),
-            RustType::Uuid => write!(f, "ID"), // TODO: Change this for actual UUID
+            RustType::Uuid => write!(f, "ID"),
             RustType::Date => write!(f, "DateTime<Utc>"),
         }
     }
